simplify the css generation logic based on context, minify the single injected css file
