+++ inc/core/styles/gutenberg.php
@@ -1,0 +1,261 @@
+<?php
+/**
+ * Style generator based on settings.
+ *
+ * @package Neve\Core\Styles
+ */
+
+namespace Neve\Core\Styles;
+
+use Neve\Core\Settings\Config;
+use Neve\Core\Settings\Mods;
+use PHPUnit\Util\Type;
+
+/**
+ * Class Generator for Gutenberg editor.
+ *
+ * @package Neve\Core\Styles
+ */
+class Gutenberg extends Generator {
+	/**
+	 * Generator constructor.
+	 */
+	public function __construct() {
+		$this->context        = Dynamic_Selector::CONTEXT_GUTENBERG;
+		$this->_subscribers[] = [
+			Dynamic_Selector::KEY_RULES    => [
+				Config::CSS_PROP_COLOR => Config::MODS_TEXT_COLOR,
+			],
+			Dynamic_Selector::KEY_SELECTOR => '
+				 .wp-block ,
+				 .editor-post-title__block .editor-post-title__input,
+				 h1,
+				 h2,
+				 h3,
+				 h4,
+				 h5,
+				 h6',
+			Dynamic_Selector::KEY_CONTEXT  => [
+				Dynamic_Selector::CONTEXT_GUTENBERG => true,
+			],
+		];
+		$this->_subscribers[] = [
+			Dynamic_Selector::KEY_RULES    => [
+				Config::CSS_PROP_BACKGROUND_COLOR => Config::MODS_BACKGROUND_COLOR,
+			],
+			Dynamic_Selector::KEY_SELECTOR => ' > *',
+			Dynamic_Selector::KEY_CONTEXT  => [
+				Dynamic_Selector::CONTEXT_GUTENBERG => true,
+			],
+		];
+		$this->_subscribers[] = [
+			Dynamic_Selector::KEY_RULES    => [
+				Config::CSS_PROP_COLOR => Config::MODS_LINK_COLOR,
+			],
+			Dynamic_Selector::KEY_SELECTOR => 'a',
+			Dynamic_Selector::KEY_CONTEXT  => [
+				Dynamic_Selector::CONTEXT_GUTENBERG => true,
+			],
+		];
+		$this->_subscribers[] = [
+			Dynamic_Selector::KEY_RULES    => [
+				Config::CSS_PROP_COLOR => Config::MODS_LINK_HOVER_COLOR,
+			],
+			Dynamic_Selector::KEY_SELECTOR => 'a:hover',
+			Dynamic_Selector::KEY_CONTEXT  => [
+				Dynamic_Selector::CONTEXT_GUTENBERG => true,
+			],
+		];
+		$this->setup_buttons();
+		$this->setup_typography();
+
+	}
+
+
+	/**
+	 * Setup typography subscribers.
+	 */
+	public function setup_typography() {
+
+
+		// Gutenberg Typography.
+
+		$this->_subscribers[] = [
+			Dynamic_Selector::KEY_SELECTOR => '.editor-post-title__block .editor-post-title__input,
+			h1,
+			h2,
+			h3,
+			h4,
+			h5,
+			h6',
+			Dynamic_Selector::KEY_RULES    => [
+				Config::CSS_PROP_FONT_FAMILY => Config::MODS_FONT_HEADINGS,
+			],
+			Dynamic_Selector::KEY_CONTEXT  => [
+				Dynamic_Selector::CONTEXT_GUTENBERG => true,
+			],
+		];
+		$this->_subscribers[] = [
+			Dynamic_Selector::KEY_SELECTOR => '#editor .editor-styles-wrapper',
+			Dynamic_Selector::KEY_RULES    => [
+				Config::CSS_PROP_FONT_FAMILY => Config::MODS_FONT_GENERAL,
+			],
+			Dynamic_Selector::KEY_CONTEXT  => [
+				Dynamic_Selector::CONTEXT_GUTENBERG => true,
+			],
+		];
+
+		$this->_subscribers[] = [
+			Dynamic_Selector::KEY_SELECTOR => '  .wp-block,
+			 [data-type="core/paragraph"] p',
+			Dynamic_Selector::KEY_RULES    => [
+				Config::CSS_PROP_FONT_SIZE      => [
+					Dynamic_Selector::META_KEY           => Config::MODS_TYPEFACE_GENERAL . '.fontSize',
+					Dynamic_Selector::META_IS_RESPONSIVE => true,
+					Dynamic_Selector::META_SUFFIX        => 'px',
+				],
+				Config::CSS_PROP_LINE_HEIGHT    => [
+					Dynamic_Selector::META_KEY           => Config::MODS_TYPEFACE_GENERAL . '.lineHeight',
+					Dynamic_Selector::META_IS_RESPONSIVE => true,
+					Dynamic_Selector::META_SUFFIX        => '',
+				],
+				Config::CSS_PROP_LETTER_SPACING => [
+					Dynamic_Selector::META_KEY           => Config::MODS_TYPEFACE_GENERAL . '.letterSpacing',
+					Dynamic_Selector::META_IS_RESPONSIVE => true,
+				],
+				Config::CSS_PROP_FONT_WEIGHT    => [
+					Dynamic_Selector::META_KEY => Config::MODS_TYPEFACE_GENERAL . '.fontWeight',
+					'font'                     => 'mods_' . Config::MODS_FONT_GENERAL,
+				],
+				Config::CSS_PROP_TEXT_TRANSFORM => Config::MODS_TYPEFACE_GENERAL . '.textTransform',
+				Config::CSS_PROP_FONT_FAMILY    => Config::MODS_FONT_GENERAL,
+			],
+			Dynamic_Selector::KEY_CONTEXT  => [
+				Dynamic_Selector::CONTEXT_GUTENBERG => true,
+			],
+		];
+		foreach (
+			[
+				'neve_h1_typeface_general' => '
+			 h1,
+			 .editor-post-title__block .editor-post-title__input',
+				'neve_h2_typeface_general' => ' h2',
+				'neve_h3_typeface_general' => ' h3',
+				'neve_h4_typeface_general' => ' h4',
+				'neve_h5_typeface_general' => ' h5',
+				'neve_h6_typeface_general' => ' h6',
+			] as $heading_mod => $heading_selector
+		) {
+			$this->_subscribers[] = [
+				Dynamic_Selector::KEY_RULES    => [
+					Config::CSS_PROP_FONT_SIZE      => [
+						Dynamic_Selector::META_KEY    => $heading_mod . '.fontSize',
+						Dynamic_Selector::META_IS_RESPONSIVE => true,
+						Dynamic_Selector::META_SUFFIX => 'em',
+					],
+					Config::CSS_PROP_LINE_HEIGHT    => [
+						Dynamic_Selector::META_KEY    => $heading_mod . '.lineHeight',
+						Dynamic_Selector::META_IS_RESPONSIVE => true,
+						Dynamic_Selector::META_SUFFIX => '',
+					],
+					Config::CSS_PROP_LETTER_SPACING => [
+						Dynamic_Selector::META_KEY => $heading_mod . '.letterSpacing',
+						Dynamic_Selector::META_IS_RESPONSIVE => true,
+					],
+					Config::CSS_PROP_FONT_WEIGHT    => [
+						Dynamic_Selector::META_KEY => $heading_mod . '.fontWeight',
+						'font'                     => 'mods_' . Config::MODS_FONT_HEADINGS,
+					],
+					Config::CSS_PROP_TEXT_TRANSFORM => $heading_mod . '.textTransform',
+					Config::CSS_PROP_FONT_FAMILY    => Config::MODS_FONT_HEADINGS,
+				],
+				Dynamic_Selector::KEY_SELECTOR => $heading_selector,
+				Dynamic_Selector::KEY_CONTEXT  => [
+					Dynamic_Selector::CONTEXT_GUTENBERG => true,
+				],
+			];
+		}
+	}
+
+	/**
+	 * Setup button subscribers.
+	 */
+	public function setup_buttons() {
+
+
+		// Gutenberg
+		$this->_subscribers[]  = [
+			Dynamic_Selector::KEY_SELECTOR => '.wp-block-button.is-style-primary .wp-block-button__link',
+			Dynamic_Selector::KEY_RULES    => [
+				Config::CSS_PROP_BACKGROUND_COLOR => Config::MODS_BUTTON_PRIMARY_STYLE . '.background',
+				Config::CSS_PROP_COLOR            => Config::MODS_BUTTON_PRIMARY_STYLE . '.text',
+				Config::CSS_PROP_BORDER_RADIUS    => Config::MODS_BUTTON_PRIMARY_STYLE . '.borderRadius',
+				Config::CSS_PROP_CUSTOM_BTN_TYPE  => Config::MODS_BUTTON_PRIMARY_STYLE . '.type',
+				Config::CSS_PROP_BORDER_WIDTH     => Config::MODS_BUTTON_PRIMARY_STYLE . '.borderWidth',
+			],
+			Dynamic_Selector::KEY_CONTEXT  => [
+				Dynamic_Selector::CONTEXT_GUTENBERG => true,
+			],
+		];
+		$this->_subscribers[]  = [
+			Dynamic_Selector::KEY_SELECTOR => '.wp-block-button.is-style-primary .wp-block-button__link:hover',
+			Dynamic_Selector::KEY_RULES    => [
+				Config::CSS_PROP_BACKGROUND_COLOR => Config::MODS_BUTTON_PRIMARY_STYLE . '.backgroundHover',
+				Config::CSS_PROP_COLOR            => Config::MODS_BUTTON_PRIMARY_STYLE . '.textHover',
+			],
+			Dynamic_Selector::KEY_CONTEXT  => [
+				Dynamic_Selector::CONTEXT_GUTENBERG => true,
+			],
+		];
+		$this->_subscribers [] = [
+			Dynamic_Selector::KEY_SELECTOR => '.wp-block-button.is-style-secondary .wp-block-button__link',
+			Dynamic_Selector::KEY_RULES    => [
+				Config::CSS_PROP_BACKGROUND_COLOR => Config::MODS_BUTTON_SECONDARY_STYLE . '.background',
+				Config::CSS_PROP_COLOR            => Config::MODS_BUTTON_SECONDARY_STYLE . '.text',
+				Config::CSS_PROP_BORDER_RADIUS    => Config::MODS_BUTTON_SECONDARY_STYLE . '.borderRadius',
+				Config::CSS_PROP_CUSTOM_BTN_TYPE  => Config::MODS_BUTTON_SECONDARY_STYLE . '.type',
+				Config::CSS_PROP_BORDER_WIDTH     => Config::MODS_BUTTON_SECONDARY_STYLE . '.borderWidth',
+			],
+			Dynamic_Selector::KEY_CONTEXT  => [
+				Dynamic_Selector::CONTEXT_GUTENBERG => true,
+			],
+		];
+		$this->_subscribers[]  = [
+			Dynamic_Selector::KEY_SELECTOR => '.wp-block-button.is-style-secondary .wp-block-button__link:hover',
+			Dynamic_Selector::KEY_RULES    => [
+				Config::CSS_PROP_BACKGROUND_COLOR => Config::MODS_BUTTON_SECONDARY_STYLE . '.backgroundHover',
+				Config::CSS_PROP_COLOR            => Config::MODS_BUTTON_SECONDARY_STYLE . '.textHover',
+			],
+			Dynamic_Selector::KEY_CONTEXT  => [
+				Dynamic_Selector::CONTEXT_GUTENBERG => true,
+			],
+		];
+
+
+		$this->_subscribers[] = [
+			Dynamic_Selector::KEY_SELECTOR => '.wp-block-button.is-style-primary .wp-block-button__link',
+			Dynamic_Selector::KEY_RULES    => [
+				Config::CSS_PROP_PADDING => [
+					Dynamic_Selector::META_KEY           => Config::MODS_BUTTON_PRIMARY_PADDING,
+					Dynamic_Selector::META_IS_RESPONSIVE => true,
+				],
+			],
+			Dynamic_Selector::KEY_CONTEXT  => [
+				Dynamic_Selector::CONTEXT_GUTENBERG => true,
+			],
+		];
+		$this->_subscribers[] = [
+			Dynamic_Selector::KEY_SELECTOR => '.wp-block-button.is-style-secondary .wp-block-button__link',
+			Dynamic_Selector::KEY_RULES    => [
+				Config::CSS_PROP_PADDING => [
+					Dynamic_Selector::META_KEY           => Config::MODS_BUTTON_SECONDARY_PADDING,
+					Dynamic_Selector::META_IS_RESPONSIVE => true,
+				],
+			],
+			Dynamic_Selector::KEY_CONTEXT  => [
+				Dynamic_Selector::CONTEXT_GUTENBERG => true,
+			],
+		];
+	}
+
+}
